chore: refactor pysimplegui related files
