+++ movie_pipeline_segments_validator/adapters/pysimplegui/config/widget.py
