+++ movie_pipeline_segments_validator/adapters/pysimplegui/controllers/detector_selector_controller.py
@@ -2,11 +2,10 @@ from typing import Any, cast
 
 import PySimpleGUI as sg
 
-from movie_pipeline_segments_validator.services import detector_service
-
-from ...domain.context import SegmentValidatorContext
-from ...domain.widget import WidgetEvent, WidgetKey
-from ...views.segments_list import render_values
+from ....domain.context import SegmentValidatorContext
+from ....services import detector_service
+from ..config.widget import WidgetEvent, WidgetKey
+from ..views.segments_list import render_values
 
 
 def populate_detector_selector(window: sg.Window, _event: str, _values: dict[str, Any]):
