+++ movie_pipeline_segments_validator/adapters/pysimplegui/controllers/segments_timeline_controller.py
@@ -2,8 +2,8 @@ from typing import Any, cast
 
 import PySimpleGUI as sg
 
-from ...domain.widget import WidgetEvent, WidgetKey
-from ...services import timeline_service
+from ....services import timeline_service
+from ..config.widget import WidgetEvent, WidgetKey
 
 
 def draw_segments(window: sg.Window, _event = None, _values = None):
