+++ movie_pipeline_segments_validator/adapters/pysimplegui/views/texts.py
