+++ movie_pipeline_segments_validator/adapters/pysimplegui/views/video.py
@@ -2,8 +2,8 @@ from typing import Any
 
 import PySimpleGUI as sg
 
-from ..adapters.pysimplegui.video_controller import render_video_new_frame, rerender_video
-from ..domain.widget import WidgetEvent, WidgetKey
+from ..config.widget import WidgetEvent, WidgetKey
+from ..controllers.video_controller import render_video_new_frame, rerender_video
 
 
 def layout():
